fix_5434 Refactor jdbc source task for the timestamp mode
