+++ ohara-connector/src/test/resources/log4j.properties
@@ -15,7 +15,7 @@
 #
 
 # Set root logger level to INFO and its only appender to console.
-log4j.rootLogger=DEBUG,console
+log4j.rootLogger=INFO,console
 
 # Custom Logging levels
 # kafka workers try to instantiate all files in classpath so there are many debug message thrown by failing to instantiate class...
